requirements.txt
# --- Core App ---
streamlit>=1.36
streamlit-extras>=0.3.6
python-dotenv>=1.0.1

# --- LangChain & LangGraph Ecosystem ---
langchain>=1.0.1
langchain-core>=1.0.1
langchain-community>=1.0.1
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.3
langgraph>=0.2.21

# --- Vectorization + Retrieval ---
sentence-transformers>=2.2.2
chromadb>=0.5.3
faiss-cpu>=1.8.0

# --- Datasets + Hugging Face ---
datasets>=3.0.0
huggingface-hub>=0.23.2

# --- Misc Utilities ---
tqdm>=4.66.5
numpy>=1.26.4
pandas>=2.2.2
